pdf/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
deprecation==2.1.0
Django==4.1.2
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
img2pdf==0.4.4
imgkit==1.2.2
lxml==4.9.1
packaging==21.3
pdfkit==1.0.0
pikepdf==6.2.0
Pillow==9.2.0
pyparsing==3.0.9
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
wkhtmltopdf==0.2
